refactor(auth): restructure project and migrate to SQLModel
- Restructured project layout for better modularity and maintainability
- Migrated token and user-related models from Pydantic to SQLModel for consistency
- Refactored auth functions and endpoint handlers to align with new structure
- Prepared base for upcoming security enhancements and feature expansion
