+++ app/main.py
@@ -1,11 +1,11 @@
 from fastapi import FastAPI
-from app.routers import registration, auth
-from app.models import create_db_and_tables
+from app.routers import registration, login
+from app.database.database import create_db_and_tables
 
 
 app = FastAPI()
 
-app.include_router(auth.router)
+app.include_router(login.router)
 app.include_router(registration.router)
