+++ app/models/user.py
@@ -1,5 +1,4 @@
 from sqlmodel import Field, SQLModel
-from app.database import engine
 
 #define user table
 class User(SQLModel, table=True):
@@ -11,7 +10,3 @@ class User(SQLModel, table=True):
     full_name: str | None = None
     hashed_password: str
     disabled: bool = Field(default=False)
-
-
-def create_db_and_tables():
-    SQLModel.metadata.create_all(engine)
