+++ requirements.txt
@@ -4,3 +4,5 @@ pyjwt
 "passlib[bcrypt]"
 'bcrypt<4.1.0'
 sqlmodel
+passlib
+dotenv
